+++ packages/blade/package.json
@@ -37,7 +37,7 @@
     "build": "run-s build:*",
     "build:clean": "rm -rf build",
     "build:generate-types": "tsc -p ./tsconfig-generate-types.json",
-    "build:generate-css-tokens": "FRAMEWORK=REACT rollup -c ./rollup-themes.config.js && node ./scripts/generateCSSTokens.js",
+    "build:generate-css-tokens": "FRAMEWORK=REACT rollup -c ./rollup-themes.config.js && node ./scripts/generateCssThemeVariables.js",
     "build:copy-declarations": "copyfiles -u 1 \"src/**/*.d.ts\" build/types",
     "build:react": "FRAMEWORK=REACT rollup -c && node ./scripts/generateRootImports.js",
     "build:react-native": "FRAMEWORK=REACT_NATIVE rollup -c && node ./scripts/generateRootImports.js",
